+++ src/App.tsx
@@ -20,7 +20,7 @@ import { Signup } from './pages/Signup';
 import { Diet } from './pages/Diet';
 import { DietLog } from './pages/DietLog';
 import { DietMeals } from './pages/DietMeals';
-import { DietMealNew } from './pages/DietMealNew';
+import { DietMealEditor } from './pages/DietMealEditor';
 import { DietWeekly } from './pages/DietWeekly';
 import { DietSettings } from './pages/DietSettings';
 import { DietFoodNew } from './pages/DietFoodNew';
@@ -118,7 +118,8 @@ function AppRoutes() {
       {/* Diet routes without bottom nav */}
       <Route path="/diet/log" element={<DietLog />} />
       <Route path="/diet/meals" element={<DietMeals />} />
-      <Route path="/diet/meals/new" element={<DietMealNew />} />
+      <Route path="/diet/meals/new" element={<DietMealEditor />} />
+      <Route path="/diet/meals/:id/edit" element={<DietMealEditor />} />
       <Route path="/diet/weekly" element={<DietWeekly />} />
       <Route path="/diet/settings" element={<DietSettings />} />
       <Route path="/diet/food/new" element={<DietFoodNew />} />